requirements.txt
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
pyyaml==6.0.1
numpy==1.26.2
pytest==7.4.3 